requirements.txt
aiogram~=3.2.0
aiohttp~=3.9.0
motor~=3.3.2
python-dotenv~=1.0.0
